codex问题处理 Signed-off-by: liuxiyao223 <liuxiyao223@huawei.com>

+++ utils/vcard/src/vcard_manager.cpp
@@ -227,8 +227,8 @@ std::vector<std::vector<std::shared_ptr<VCardContact>>> VCardManager::SplitConta
         std::vector<std::shared_ptr<VCardContact>>::iterator endPtr = list.end();
         std::vector<std::shared_ptr<VCardContact>>::iterator end;
         while (curPtr < endPtr) {
-            end = static_cast<size_t>(endPtr - curPtr) > step ? (step + curPtr) : endPtr;
-            step = static_cast<size_t>(endPtr - curPtr) > step ? step : (endPtr - curPtr);
+            end = static_cast<size_t>(endPtr - curPtr) > step ? static_cast<size_t>(step + curPtr) : endPtr;
+            step = static_cast<size_t>(endPtr - curPtr) > step ? step : static_cast<size_t>(endPtr - curPtr);
             result.push_back(std::vector<std::shared_ptr<VCardContact>>(curPtr, end));
             curPtr += step;
         }
